pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cicd-fixer"
version = "0.1.0"
description = "AI-powered CI/CD pipeline optimizer for GitHub Actions and GitLab CI"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "CI Optimizer Team", email = "team@cicd-fixer.dev"}
]
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Build Tools",
]
dependencies = [
    "typer[all]>=0.9.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
    "networkx>=3.0",
    "jsonschema>=4.0.0",
    "click>=8.0.0",
    "colorama>=0.4.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "flake8>=6.0.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
cicd-fixer = "cli.cli_entry:main"
ci-agent = "cli.cli_entry:main"

[tool.setuptools.packages.find]
where = ["ci-agent"]
include = ["agent*", "cli*"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["ci-agent/tests"]
pythonpath = ["ci-agent"]
addopts = "-v --cov=agent --cov=cli --cov-report=term-missing" 